Cleanup and small fixes
Update some exports, fix indentation, and move model selection to models
folder.
